Migrate gantt plugin
